Refactor implementation
