+++ opentelemetry-api/src/opentelemetry/util/_providers.py
@@ -14,10 +14,9 @@
 
 from logging import getLogger
 from os import environ
-from sys import version_info
 from typing import TYPE_CHECKING, TypeVar, cast
 
-from opentelemetry.util._entry_points import entry_points
+from opentelemetry.util._importlib_metadata import entry_points
 
 if TYPE_CHECKING:
     from opentelemetry.metrics import MeterProvider
@@ -38,14 +37,6 @@ def _load_provider(
             str,
             environ.get(provider_environment_variable, f"default_{provider}"),
         )
-
-        # FIXME remove when support for 3.9 is dropped.
-        if version_info.minor in (8, 9):
-
-            for entry_point in entry_points()[f"opentelemetry_{provider}"]:  # type: ignore
-                if entry_point.name == provider_name:  # type: ignore
-                    return cast(Provider, entry_point.load()())  # type: ignore
-            raise Exception(f"Provider {provider_name} not found")
 
         return cast(
             Provider,
